Add login page with simple session guard
