+++ app/lib/auth.ts
@@ -1,0 +1,8 @@
+export const SESSION_COOKIE_NAME = "coachvisio-session"
+
+export function areCredentialsValid(
+  identifiant: FormDataEntryValue | null,
+  motDePasse: FormDataEntryValue | null,
+): boolean {
+  return identifiant === "Test" && motDePasse === "Test"
+}
